requirements.txt
discord~=2.3.0
humanize~=4.6.0
python-dotenv~=1.0.0
pyarr~=5.1.0
thefuzz~=0.19.0